yololm/requirements.txt
accelerate==0.30.1
addict==2.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
einops==0.8.0
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
huggingface-hub==0.17.3
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mmcv-full @ file:///home/chenhui/wangao/projects/yolo-world/gpt4roi/mmcv_full-1.7.2-cp310-cp310-manylinux1_x86_64.whl#sha256=363ff27d7ed071eb209c975b55bf0c3434554769ae9be554e0a20533a2259c5d
mpmath==1.3.0
networkx==3.4.2
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
psutil==6.1.1
pycocotools==2.0.8
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sentencepiece==0.2.0
six==1.17.0
sympy==1.13.3
terminaltables==3.1.10
tokenizers==0.14.1
tomli==2.2.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.67.1
transformers==4.34.0
triton==2.1.0
typing_extensions==4.12.2
urllib3==2.3.0
yapf==0.43.0
